requirements.txt
AI Core & Image Processing
torch>=2.0.0
torchvision>=0.15.0
paddlepaddle-gpu==2.6.0 # GPU 버전 (CPU만 쓴다면 paddlepaddle)
paddleocr>=2.7.0
ultralytics>=8.0.0 # YOLOv8
opencv-python-headless>=4.8.0 # 서버용 (GUI 없음)
numpy>=1.24.0
Pillow>=10.0.0

GOT-OCR Dependencies (HuggingFace)
transformers>=4.36.0
accelerate>=0.25.0
tiktoken>=0.5.0
scipy
verovio # 악보/표 렌더링용 (GOT 권장)

Web Server & Async Queue
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
celery>=5.3.6
redis>=5.0.1

LLM API
google-generativeai>=0.3.2
python-dotenv>=1.0.0
